Fix frame rate regulation
This commit ammends some of the frame rate regulations introduced within the last commit. It corrects the comments stating that the TIME_PER_FRAME constant acts as a framerate cap. Instead it is used as a maximum permitted time for a frame render. This commit also introduces a framerate limit and implements it within the sf::RenderWindow used to display the game.