This commit adds a ViewManager and introduces the GameView
Within this commit I have introduced a blank GameView which will house
the game itself. I have also, and more significantly, added a
ViewManager which is charged with handling the changing between Views to
make it easier to switch between screens, i.e. from the Main Menu to the
Game.