Add thread synchronisation for loading screen and create thread as class member
This commit introduces thread synchronisation in the loading screen
class using std::atomic<> to eliminate the potential for a race
condition.

This commit also introduces the thread declaration as part of the
LoadingScreen class instead of within the LoadingScreen::run() method
which would end up prematurely terminating the thread as the thread
would be destroyed when the function finished.